Extract utility for reuse
